move empty to empty.go
